refactor:#8 MemberService 모듈이동
